Did some reorganization of packages/classes
