+++ src/nmt/minecraft/QuestManager/Quest/Quest.java
@@ -11,11 +11,12 @@ import java.util.Set;
 
 import nmt.minecraft.QuestManager.QuestManager;
 import nmt.minecraft.QuestManager.QuestManagerPlugin;
-import nmt.minecraft.QuestManager.Configuration.GoalState;
-import nmt.minecraft.QuestManager.Configuration.QuestState;
+import nmt.minecraft.QuestManager.Configuration.State.GoalState;
+import nmt.minecraft.QuestManager.Configuration.State.QuestState;
 import nmt.minecraft.QuestManager.Player.QuestPlayer;
 import nmt.minecraft.QuestManager.Quest.History.History;
 import nmt.minecraft.QuestManager.Quest.Requirements.Requirement;
+import nmt.minecraft.QuestManager.Quest.Requirements.RequirementUpdateEvent;
 
 import org.bukkit.configuration.InvalidConfigurationException;
 import org.bukkit.event.EventHandler;
